// src/generated/prisma/enums.ts
export const MetalType = {
  GOLD_9K: 'GOLD_9K',
  GOLD_14K: 'GOLD_14K',
  GOLD_18K: 'GOLD_18K',
  GOLD_22K: 'GOLD_22K',
  GOLD_24K: 'GOLD_24K',
  PLATINUM: 'PLATINUM',
  PALLADIUM: 'PALLADIUM',
  STERLING_SILVER: 'STERLING_SILVER',
  FINE_SILVER: 'FINE_SILVER',
  TITANIUM: 'TITANIUM',
  MIXED_METALS: 'MIXED_METALS',
  OTHER: 'OTHER'
} as const

export type MetalType = (typeof MetalType)[keyof typeof MetalType]


export const MetalColor = {
  YELLOW: 'YELLOW',
  WHITE: 'WHITE',
  ROSE: 'ROSE',
  TWO_TONE: 'TWO_TONE',
  NATURAL: 'NATURAL'
} as const

export type MetalColor = (typeof MetalColor)[keyof typeof MetalColor]


export const Gemstone = {
  NONE: 'NONE',
  DIAMOND: 'DIAMOND',
  LAB_GROWN_DIAMOND: 'LAB_GROWN_DIAMOND',
  SAPPHIRE: 'SAPPHIRE',
  RUBY: 'RUBY',
  EMERALD: 'EMERALD',
  PEARL: 'PEARL',
  MOISSANITE: 'MOISSANITE',
  TANZANITE: 'TANZANITE',
  AQUAMARINE: 'AQUAMARINE',
  MORGANITE: 'MORGANITE',
  OPAL: 'OPAL',
  AMETHYST: 'AMETHYST',
  TOPAZ: 'TOPAZ',
  TOURMALINE: 'TOURMALINE',
  SPINEL: 'SPINEL',
  GARNET: 'GARNET',
  CITRINE: 'CITRINE',
  PERIDOT: 'PERIDOT',
  ONYX: 'ONYX',
  TURQUOISE: 'TURQUOISE',
  JADE: 'JADE',
  OTHER: 'OTHER'
} as const

export type Gemstone = (typeof Gemstone)[keyof typeof Gemstone]


export const StoneShape = {
  ROUND: 'ROUND',
  OVAL: 'OVAL',
  CUSHION: 'CUSHION',
  PRINCESS: 'PRINCESS',
  EMERALD: 'EMERALD',
  PEAR: 'PEAR',
  MARQUISE: 'MARQUISE',
  RADIANT: 'RADIANT',
  ASSCHER: 'ASSCHER',
  HEART: 'HEART',
  BAGUETTE: 'BAGUETTE',
  CABOCHON: 'CABOCHON',
  ROSE_CUT: 'ROSE_CUT',
  OLD_EUROPEAN: 'OLD_EUROPEAN',
  OTHER: 'OTHER'
} as const

export type StoneShape = (typeof StoneShape)[keyof typeof StoneShape]


export const ClarityGrade = {
  FL: 'FL',
  IF: 'IF',
  VVS1: 'VVS1',
  VVS2: 'VVS2',
  VS1: 'VS1',
  VS2: 'VS2',
  SI1: 'SI1',
  SI2: 'SI2',
  I1: 'I1',
  I2: 'I2',
  I3: 'I3'
} as const

export type ClarityGrade = (typeof ClarityGrade)[keyof typeof ClarityGrade]


export const CutGrade = {
  EXCELLENT: 'EXCELLENT',
  VERY_GOOD: 'VERY_GOOD',
  GOOD: 'GOOD',
  FAIR: 'FAIR',
  POOR: 'POOR'
} as const

export type CutGrade = (typeof CutGrade)[keyof typeof CutGrade]


export const CertificateLab = {
  GIA: 'GIA',
  IGI: 'IGI',
  HRD: 'HRD',
  AGS: 'AGS',
  GCAL: 'GCAL',
  SSEF: 'SSEF',
  GUBELIN: 'GUBELIN',
  AGL: 'AGL',
  BIS_HALLMARK: 'BIS_HALLMARK',
  ASSAY_OFFICE: 'ASSAY_OFFICE',
  OTHER: 'OTHER'
} as const

export type CertificateLab = (typeof CertificateLab)[keyof typeof CertificateLab]
// ...
export const SpotMetal = {
  GOLD: 'GOLD',
  SILVER: 'SILVER',
  PLATINUM: 'PLATINUM',
  PALLADIUM: 'PALLADIUM'
} as const

export type SpotMetal = (typeof SpotMetal)[keyof typeof SpotMetal]